requirements.txt
flask==3.0.3            
connexion==3.1.0    
flask-testing==0.8.1    
pytest==7.1.2           
requests==2.28.1        
python-dotenv==1.0.1
flasgger>=0.9.7
Flask-CORS==5.0.0